+++ lib/generators/solid_queue_guard/install/templates/solid_queue_guard.rb
@@ -23,6 +23,7 @@ SolidQueueGuard.configure do |config|
   # config.health_token = ENV["SOLID_QUEUE_GUARD_TOKEN"]
   # config.integrate_rails_health = true
   # config.integrate_mission_control = true
+  # config.on_status_change = ->(previous, current, report) { ... }
   # config.notify_with = [:rails_logger, :slack, :datadog, :webhook]
   # config.metrics_backends = [:statsd, :prometheus, :opentelemetry]
 end