+++ lib/solid_queue_guard/mission_control/integration.rb
@@ -2,6 +2,7 @@
 
 module SolidQueueGuard
   module MissionControl
+    # @api private
     class Integration
       class << self
         def install!