+++ lib/solid_queue_guard/mission_control/navigation_extension.rb
@@ -2,6 +2,7 @@
 
 module SolidQueueGuard
   module MissionControl
+    # @api private
     module NavigationExtension
       def navigation_sections
         super.tap do |sections|